+++ lib/active_attr/matchers/have_attribute_matcher.rb
@@ -25,24 +25,6 @@ module ActiveAttr
       attr_reader :attribute_name, :default_value
       private :attribute_name, :default_value
 
-      # Specify that the attribute should have the given default value
-      #
-      # @example Person's first name should default to John
-      #   describe Person do
-      #     it do
-      #       should have_attribute(:first_name).with_default_value_of("John")
-      #     end
-      #   end
-      #
-      # @param [Object]
-      #
-      # @since 0.5.0
-      def with_default_value_of(default_value)
-        @default_value_set = true
-        @default_value = default_value
-        self
-      end
-
       # @return [String] Description
       # @private
       def description
@@ -56,6 +38,8 @@ module ActiveAttr
           "expected #{@model_class.name} to include ActiveAttr::Attributes"
         elsif !includes_defaults?
           "expected #{@model_class.name} to include ActiveAttr::AttributeDefaults"
+        elsif !includes_typecasting?
+          "expected #{@model_class.name} to include ActiveAttr::TypecastedAttributes"
         else
           "expected #{@model_class.name} to have #{attribute_description}"
         end
@@ -67,17 +51,35 @@ module ActiveAttr
         raise TypeError, "can't convert #{attribute_name.class} into Symbol" unless attribute_name.respond_to? :to_sym
         @attribute_name = attribute_name.to_sym
         @default_value_set = false
+        @type = nil
+      end
+
+      # Specify that the attribute should have the given type
+      #
+      # @example Person's first name should be a String
+      #   describe Person do
+      #     it { should have_attribute(:first_name).of_type(String) }
+      #   end
+      #
+      # @param [Class] type The expected type
+      #
+      # @return [HaveAttributeMatcher] The matcher
+      #
+      # @since 0.5.0
+      def of_type(type)
+        @type = type
+        self
       end
 
       # @private
       def matches?(model_or_model_class)
         @model_class = class_from(model_or_model_class)
 
-        return false if !includes_attributes? || !includes_defaults?
+        return false if !includes_attributes? || !includes_defaults? || !includes_typecasting?
 
         @attribute_definition = @model_class.attributes[attribute_name]
 
-        !!(@attribute_definition && (!@default_value_set || @attribute_definition[:default] == default_value))
+        !!(@attribute_definition && default_matches? && type_matches?)
       end
 
       # @return [String] Negative failure message
@@ -86,22 +88,61 @@ module ActiveAttr
         "expected #{@model_class.name} to not have #{attribute_description}"
       end
 
+      # Specify that the attribute should have the given default value
+      #
+      # @example Person's first name should default to John
+      #   describe Person do
+      #     it do
+      #       should have_attribute(:first_name).with_default_value_of("John")
+      #     end
+      #   end
+      #
+      # @param [Object] default_value The expected default value
+      #
+      # @return [HaveAttributeMatcher] The matcher
+      #
+      # @since 0.5.0
+      def with_default_value_of(default_value)
+        @default_value_set = true
+        @default_value = default_value
+        self
+      end
+
       private
 
       def attribute_description
-        "attribute named #{attribute_name}#{ " with a default value of #{default_value.inspect}" if @default_value_set}"
+        "attribute named #{attribute_name}".tap do |result|
+          result << " of type #{@type}" if @type
+          result << " with a default value of #{default_value.inspect}" if @default_value_set
+        end
       end
 
       def includes_attributes?
-        @model_class.ancestors.map(&:name).include?("ActiveAttr::Attributes")
+        model_ancestor_names.include?("ActiveAttr::Attributes")
       end
 
       def includes_defaults?
-        !@default_value_set || @model_class.ancestors.map(&:name).include?("ActiveAttr::AttributeDefaults")
+        !@default_value_set || model_ancestor_names.include?("ActiveAttr::AttributeDefaults")
+      end
+
+      def includes_typecasting?
+        !@type || model_ancestor_names.include?("ActiveAttr::TypecastedAttributes")
+      end
+
+      def model_ancestor_names
+        @model_class.ancestors.map(&:name)
       end
 
       def class_from(object)
         Class === object ? object : object.class
+      end
+
+      def default_matches?
+        !@default_value_set || @attribute_definition[:default] == default_value
+      end
+
+      def type_matches?
+        !@type || @model_class._attribute_type(attribute_name) == @type
       end
     end
   end