+++ spec/functional/active_attr/matchers/have_attribute_matcher_spec.rb
@@ -2,6 +2,7 @@ require "spec_helper"
 require "active_attr/attributes"
 require "active_attr/attribute_defaults"
 require "active_attr/matchers/have_attribute_matcher"
+require "active_attr/typecasted_attributes"
 
 module ActiveAttr
   module Matchers
@@ -10,6 +11,7 @@ module ActiveAttr
         Class.new do
           include Attributes
           include AttributeDefaults
+          include TypecastedAttributes
 
           def self.name
             "Person"
@@ -56,6 +58,28 @@ module ActiveAttr
           before { subject.matches?(model_class) }
 
           it { subject.failure_message.should == %{expected #{model_class.name} to include ActiveAttr::AttributeDefaults} }
+        end
+      end
+
+      shared_examples "a matcher matching a class without ActiveAttr::TypecastedAttributes" do
+        let :model_class do
+          Class.new do
+            include Attributes
+
+            def self.name
+              "Person"
+            end
+          end
+        end
+
+        describe "#matches?" do
+          it { subject.matches?(model_class).should be_false }
+        end
+
+        describe "#failure_message" do
+          before { subject.matches?(model_class) }
+
+          it { subject.failure_message.should == %{expected #{model_class.name} to include ActiveAttr::TypecastedAttributes} }
         end
       end
 
@@ -234,6 +258,104 @@ module ActiveAttr
             before { subject.matches?(model_class) }
 
             it { subject.negative_failure_message.should == %{expected Person to not have attribute named first_name with a default value of nil} }
+          end
+        end
+      end
+
+      context "a matcher with a type" do
+        subject { described_class.new(:first_name).of_type(String) }
+
+        it_should_behave_like "a matcher matching a class without ActiveAttr::Attributes"
+        it_should_behave_like "a matcher matching a class without ActiveAttr::TypecastedAttributes"
+
+        context "a class with the attribute but no type" do
+          before { model_class.attribute :first_name }
+
+          describe "#matches?" do
+            it { subject.matches?(model_class).should be_false }
+          end
+
+          describe "#failure_message" do
+            before { subject.matches?(model_class) }
+
+            it { subject.failure_message.should == %{expected Person to have attribute named first_name of type String} }
+          end
+        end
+
+        context "a class with the attribute and a different type" do
+          before { model_class.attribute :first_name, :type => Symbol }
+
+          describe "#matches?" do
+            it { subject.matches?(model_class).should be_false }
+          end
+
+          describe "#failure_message" do
+            before { subject.matches?(model_class) }
+
+            it { subject.failure_message.should == %{expected Person to have attribute named first_name of type String} }
+          end
+        end
+
+        context "a class with the attribute and the right type" do
+          before { model_class.attribute :first_name, :type => String }
+
+          describe "#matches?" do
+            it { subject.matches?(model_class).should be_true }
+          end
+
+          describe "#negative_failure_message" do
+            before { subject.matches?(model_class) }
+
+            it { subject.negative_failure_message.should == %{expected Person to not have attribute named first_name of type String} }
+          end
+        end
+      end
+
+      context "a matcher with a type of Object" do
+        subject { described_class.new(:first_name).of_type(Object) }
+
+        it_should_behave_like "a matcher matching a class without ActiveAttr::Attributes"
+        it_should_behave_like "a matcher matching a class without ActiveAttr::TypecastedAttributes"
+
+        context "a class with the attribute but no type" do
+          before { model_class.attribute :first_name }
+
+          describe "#matches?" do
+            it { subject.matches?(model_class).should be_true }
+          end
+
+          describe "#negative_failure_message" do
+            before { subject.matches?(model_class) }
+
+            it { subject.negative_failure_message.should == %{expected Person to not have attribute named first_name of type Object} }
+          end
+        end
+
+        context "a class with the attribute and a different type" do
+          before { model_class.attribute :first_name, :type => String }
+
+          describe "#matches?" do
+            it { subject.matches?(model_class).should be_false }
+          end
+
+          describe "#failure_message" do
+            before { subject.matches?(model_class) }
+
+            it { subject.failure_message.should == %{expected Person to have attribute named first_name of type Object} }
+          end
+        end
+
+        context "a class with the attribute and the right type" do
+          before { model_class.attribute :first_name, :type => Object }
+
+          describe "#matches?" do
+            it { subject.matches?(model_class).should be_true }
+          end
+
+          describe "#negative_failure_message" do
+            before { subject.matches?(model_class) }
+
+            it { subject.negative_failure_message.should == %{expected Person to not have attribute named first_name of type Object} }
           end
         end
       end